% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurkinaFarms_polys.R
\docType{data}
\name{BurkinaFarms_polys}
\alias{BurkinaFarms_polys}
\title{BurkinaFarms_polys}
\format{
This file contains the polygons of the regions of Burkina Faso.
}
\source{
<https://data.humdata.org/dataset/burkina-faso-administrative-boundaries>
}
\usage{
BurkinaFarms_polys
}
\description{
Dataset of Burkina Faso - Subnational Administrative Boundaries
}
\keyword{datasets}
