% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_mgcv.R
\name{cop}
\alias{cop}
\title{cop family}
\usage{
cop(link = list("tanh"), u1, u2, cop_family = 1)
}
\arguments{
\item{link}{specifying the link for the \eqn{\tau}.}

\item{u1}{probability integral transformed observations for margin 1.}

\item{u2}{probability integral transformed observations for margin 2.}

\item{cop_family}{integer, defines the copula family:\cr
\code{1} = Gaussian copula \cr}
}
\value{
An object inheriting from class \code{general.family} of the mgcv package, which can be used in the dsfa package.
}
\description{
The cop family implements multiple copula distributions in which the \eqn{\tau} can depend on additive predictors. Useable only with \code{mgcv::gam}, the additive predictors is specified via a formulae.
}
\details{
Internal function. Used with gam to fit copula model, which in turn is used for starting values. The function \code{gam} is from the mgcv package is called with a formulae.
}
\keyword{internal}
